+++ cmd/kubectl-package/command/rolloutcmd/rollout_utils.go
@@ -1,4 +1,4 @@
-package rolloututils
+package rolloutcmd
 
 import (
 	"context"
